refactor(app_runtime): Memory handler を memory.rs に抽出 (SPEC-2077 Phase B)
- crates/gwt/src/app_runtime/memory.rs を新規作成し、load_memo_events / create_memo_note_events / update_memo_note_events / delete_memo_note_events と private helper (resolve_memo_window_context / memo_window_ids_for_tab / memo_snapshot_events) を移行
- mod.rs に mod memory; 宣言を追加 (board の隣)
- god module 解体 phase chain の Phase B、 Phase A pattern を reuse して 1 module 抽出 + dispatch 経路は impl 自動 routing で 0 変更

Verification:
- cargo test -p gwt-core -p gwt: 全 pass
- cargo clippy -- -D warnings: 0 warnings
- cargo fmt --check: 差分なし
- mod.rs 7293→7190 行 (Phase B で 103 行減、 累計 255 行減)
- memory.rs 新 253 行 (memo handler の単一責務 module)
