+++ crates/gwt/src/app_runtime/mod.rs
@@ -143,6 +143,7 @@ pub(crate) type AgentLaunchCompletion = (
 pub(crate) type AgentLaunchResult = Result<AgentLaunchCompletion, String>;
 
 mod board;
+mod memory;
 pub(crate) use board::BoardPostRequest;
 
 fn dispatch_agent_launch_success<F>(
@@ -1710,40 +1711,6 @@ impl AppRuntime {
         }
     }
 
-    pub(crate) fn load_memo_events(&self, client_id: &str, id: &str) -> Vec<OutboundEvent> {
-        let project_root = match self.resolve_memo_window_context(id) {
-            Ok(context) => context,
-            Err(message) => {
-                return vec![OutboundEvent::reply(
-                    client_id,
-                    BackendEvent::MemoError {
-                        id: id.to_string(),
-                        message,
-                    },
-                )];
-            }
-        }
-        .1;
-
-        match gwt_core::notes::load_snapshot(&project_root) {
-            Ok(snapshot) => vec![OutboundEvent::reply(
-                client_id,
-                BackendEvent::MemoNotes {
-                    id: id.to_string(),
-                    notes: snapshot.notes,
-                    selected_note_id: None,
-                },
-            )],
-            Err(error) => vec![OutboundEvent::reply(
-                client_id,
-                BackendEvent::MemoError {
-                    id: id.to_string(),
-                    message: error.to_string(),
-                },
-            )],
-        }
-    }
-
     pub(crate) fn load_logs_events(&self, client_id: &str, id: &str) -> Vec<OutboundEvent> {
         let Some(address) = self.window_lookup.get(id) else {
             return vec![OutboundEvent::reply(
@@ -2106,203 +2073,6 @@ impl AppRuntime {
         }
 
         events
-    }
-
-    pub(crate) fn create_memo_note_events(
-        &self,
-        client_id: &str,
-        id: &str,
-        title: String,
-        body: String,
-        pinned: bool,
-    ) -> Vec<OutboundEvent> {
-        let (tab_id, project_root) = match self.resolve_memo_window_context(id) {
-            Ok(context) => context,
-            Err(message) => {
-                return vec![OutboundEvent::reply(
-                    client_id,
-                    BackendEvent::MemoError {
-                        id: id.to_string(),
-                        message,
-                    },
-                )];
-            }
-        };
-
-        let created = match gwt_core::notes::create_note(
-            &project_root,
-            gwt_core::notes::MemoNoteDraft::new(title, body, pinned),
-        ) {
-            Ok(note) => note,
-            Err(error) => {
-                return vec![OutboundEvent::reply(
-                    client_id,
-                    BackendEvent::MemoError {
-                        id: id.to_string(),
-                        message: error.to_string(),
-                    },
-                )];
-            }
-        };
-
-        self.memo_snapshot_events(&tab_id, id, Some(created.id), &project_root, client_id)
-    }
-
-    pub(crate) fn update_memo_note_events(
-        &self,
-        client_id: &str,
-        id: &str,
-        note_id: &str,
-        title: String,
-        body: String,
-        pinned: bool,
-    ) -> Vec<OutboundEvent> {
-        let (tab_id, project_root) = match self.resolve_memo_window_context(id) {
-            Ok(context) => context,
-            Err(message) => {
-                return vec![OutboundEvent::reply(
-                    client_id,
-                    BackendEvent::MemoError {
-                        id: id.to_string(),
-                        message,
-                    },
-                )];
-            }
-        };
-
-        let updated = match gwt_core::notes::update_note(
-            &project_root,
-            note_id,
-            gwt_core::notes::MemoNoteDraft::new(title, body, pinned),
-        ) {
-            Ok(note) => note,
-            Err(error) => {
-                return vec![OutboundEvent::reply(
-                    client_id,
-                    BackendEvent::MemoError {
-                        id: id.to_string(),
-                        message: error.to_string(),
-                    },
-                )];
-            }
-        };
-
-        self.memo_snapshot_events(&tab_id, id, Some(updated.id), &project_root, client_id)
-    }
-
-    pub(crate) fn delete_memo_note_events(
-        &self,
-        client_id: &str,
-        id: &str,
-        note_id: &str,
-    ) -> Vec<OutboundEvent> {
-        let (tab_id, project_root) = match self.resolve_memo_window_context(id) {
-            Ok(context) => context,
-            Err(message) => {
-                return vec![OutboundEvent::reply(
-                    client_id,
-                    BackendEvent::MemoError {
-                        id: id.to_string(),
-                        message,
-                    },
-                )];
-            }
-        };
-
-        if let Err(error) = gwt_core::notes::delete_note(&project_root, note_id) {
-            return vec![OutboundEvent::reply(
-                client_id,
-                BackendEvent::MemoError {
-                    id: id.to_string(),
-                    message: error.to_string(),
-                },
-            )];
-        }
-
-        let selected_note_id = match gwt_core::notes::load_snapshot(&project_root) {
-            Ok(snapshot) => snapshot.notes.first().map(|note| note.id.clone()),
-            Err(error) => {
-                return vec![OutboundEvent::reply(
-                    client_id,
-                    BackendEvent::MemoError {
-                        id: id.to_string(),
-                        message: error.to_string(),
-                    },
-                )];
-            }
-        };
-
-        self.memo_snapshot_events(&tab_id, id, selected_note_id, &project_root, client_id)
-    }
-
-    fn resolve_memo_window_context(
-        &self,
-        id: &str,
-    ) -> std::result::Result<(String, PathBuf), String> {
-        let Some(address) = self.window_lookup.get(id) else {
-            return Err("Window not found".to_string());
-        };
-        let Some(tab) = self.tab(&address.tab_id) else {
-            return Err("Project tab not found".to_string());
-        };
-        let Some(window) = tab.workspace.window(&address.raw_id) else {
-            return Err("Window not found".to_string());
-        };
-        if window.preset != WindowPreset::Memo {
-            return Err("Window is not a Memo surface".to_string());
-        }
-
-        Ok((address.tab_id.clone(), tab.project_root.clone()))
-    }
-
-    fn memo_window_ids_for_tab(&self, tab_id: &str) -> Vec<String> {
-        let Some(tab) = self.tab(tab_id) else {
-            return Vec::new();
-        };
-        tab.workspace
-            .persisted()
-            .windows
-            .iter()
-            .filter(|window| window.preset == WindowPreset::Memo)
-            .map(|window| combined_window_id(tab_id, &window.id))
-            .collect()
-    }
-
-    fn memo_snapshot_events(
-        &self,
-        tab_id: &str,
-        selected_window_id: &str,
-        selected_note_id: Option<String>,
-        project_root: &Path,
-        client_id: &str,
-    ) -> Vec<OutboundEvent> {
-        let snapshot = match gwt_core::notes::load_snapshot(project_root) {
-            Ok(snapshot) => snapshot,
-            Err(error) => {
-                return vec![OutboundEvent::reply(
-                    client_id,
-                    BackendEvent::MemoError {
-                        id: selected_window_id.to_string(),
-                        message: error.to_string(),
-                    },
-                )];
-            }
-        };
-
-        self.memo_window_ids_for_tab(tab_id)
-            .into_iter()
-            .map(|window_id| {
-                OutboundEvent::broadcast(BackendEvent::MemoNotes {
-                    id: window_id.clone(),
-                    notes: snapshot.notes.clone(),
-                    selected_note_id: if window_id == selected_window_id {
-                        selected_note_id.clone()
-                    } else {
-                        None
-                    },
-                })
-            })
-            .collect()
     }
 
     pub(crate) fn handle_board_projection_changed_events(
